Made a few changes to the audio API.

+++ nitro/src/component/mod.rs
@@ -25,6 +25,6 @@ macro_rules! default_component_impl {
 
 pub enum Message {
     Start { key: i32 },
-    Update { delta_time: f64 },
+    Update { delta_time: f32 },
     UserMessage(Box<Any>),
 }
